* Allow notification of failed bundles * Move already installed bundles into the failed directory * Moved triage functions outside of scanner into a separate module
